Refactor: Split layers (entiry/view/service/dto)
- delete unused assets
- Sync coding convention (variable names as camel case)
