+++ results/hojongs/src/entry/user.ts
@@ -1,8 +1,8 @@
 import Vue from 'vue'
-import Login from './Login.vue'
+import User from '@/view/User.vue'
 
 Vue.config.productionTip = false
 
 new Vue({
-  render: h => h(Login),
+  render: h => h(User),
 }).$mount('#app')
