+++ results/hojongs/vue.config.js
@@ -1,7 +1,7 @@
 module.exports = {
     // https://cli.vuejs.org/config/#pages
     pages: {
-        index: 'src/main.ts',
-        user: 'src/user.ts',
+        index: 'src/entry/main.ts',
+        user: 'src/entry/user.ts',
     }
 }
